Organized Battle System And Terrain generation Component
renamed a couple of scripts and created some new folders in a way that shouldn't cause conflicts with the code of others but made more logical sense.  added more functionality to the base classes and improved the animators of all characters such that all string names that trigger animations will be the same for ALL characters.